+++ src/js/fake-translator.js
@@ -27,7 +27,11 @@ var SOURCE_STRINGS = {
         'placeholder': 'Click here to upload file. (< __maxSize__)',
         'notFound': 'File __existing__ could not be found (leave unchanged if already submitted and you want to preserve it).',
         'waitingForPermissions': 'Waiting for user permissions.',
-        'resetWarning': 'This will remove the file. Are you sure you want to do this?'
+        'resetWarning': 'This will remove the file. Are you sure you want to do this?',
+        'toolargeerror': 'File too large (> __maxSize__)'
+    },
+    'drawwidget': {
+        'resetWarning': 'This will remove the drawing, signature and/or file. Are you sure you want to do this?'
     },
     'form': {
         'required': 'required'