+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Full support for annotate widget.
+
+##### Changed
+- Draw/signature widget requires confirmation before reset.
+
 [4.45.0] - 2018-02-08
 ---------------------
 ##### Added